+++ src/api/tracking.ts
@@ -1,0 +1,9 @@
+import { api } from ".";
+
+// background fetch
+export const fetchTrackings = async (params: {
+  orderId: string;
+  latitude: number;
+  longitude: number;
+  trackingTime: string;
+}) => await api.post("track", { json: params });
